enums moved to entities
