+++ conveyor/src/main/java/com/aryunin/conveyor/entity/enums/MaterialStatus.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.util;
+package com.aryunin.conveyor.entity.enums;
 
 public enum MaterialStatus {
     POOR,
