+++ conveyor/src/main/java/com/aryunin/conveyor/entity/enums/Position.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.util;
+package com.aryunin.conveyor.entity.enums;
 
 public enum Position {
     WORKER,
